+++ mono/mind_maps/static/mind_maps/style.js
@@ -1,0 +1,28 @@
+const STYLES = {
+    0: {
+        fontSize: 2.5, padding: 3, borderSize: .8,
+        textStyle: {
+            bold: true, italic: false, underline: true, lineThrough: false
+        }
+    },
+    1: {
+        fontSize: 1.5, padding: 2, borderSize: .6,
+        textStyle: {
+            bold: true, italic: false, underline: false, lineThrough: false
+        }
+    },
+    2: {
+        fontSize: 1.5, padding: 2, borderSize: .4,
+        textStyle: {
+            bold: false, italic: true, underline: false, lineThrough: false
+        }
+    },
+}
+
+const PRESET_COLORS = {
+    'red': {
+        background: '#FFCCCC',
+        border: '#DB2828',
+        font: '#DB2828',
+    },
+}